Fix #256 Use async process to query Device API
